Criadas funções de edição e de exclusão na controller de categoria
